refactor(diffusers): move the SD ControlNet under models/controlnets
Mirrors the upstream diffusers layout: diffusers moved every controlnet into
models/controlnets/ in #8768 (first released in v0.32.0; the old
models/controlnet.py deprecation shim is gone as of current releases). This
also lines the models/ tree up with the existing autoencoders/, transformers/
and unets/ subpackages.

Pure move (git tracks it as a rename) plus import-path updates; no behavior
change. TestSDControlNetModel / TestSDMultiControlNetModel pass on NPU.
