+++ src/optimum/rbln/diffusers/models/controlnets/controlnet.py
@@ -19,12 +19,12 @@ from diffusers import ControlNetModel
 from diffusers.models.controlnets.controlnet import ControlNetOutput
 from transformers import PretrainedConfig
 
-from ...configuration_utils import RBLNCompileConfig, RBLNModelConfig
-from ...modeling import RBLNModel
-from ...utils.logging import get_logger
-from ...utils.model_utils import get_rbln_model_cls
-from ..configurations import RBLNControlNetModelConfig
-from ..modeling_diffusers import RBLNDiffusionMixin, RBLNDiffusionMixinConfig
+from ....configuration_utils import RBLNCompileConfig, RBLNModelConfig
+from ....modeling import RBLNModel
+from ....utils.logging import get_logger
+from ....utils.model_utils import get_rbln_model_cls
+from ...configurations import RBLNControlNetModelConfig
+from ...modeling_diffusers import RBLNDiffusionMixin, RBLNDiffusionMixinConfig
 
 
 if TYPE_CHECKING:
